Implement open shop logic
Implemented base UI to test opening and closing shop on E input. Paused the game when the shop is open. Experienced issue with being able to open shop while in pause or inventory menu. Fixed to only open the shop menu if the game is not paused as the game is paused in these other screens.